Ruby on Travis CI is hard
Running an Ansible playbook on Travis CI that requires a version of Ruby
newer than the system version of Ruby is tricky. As was discovered in
923ec94, the system version of Ruby is used by gitsh's installer unless
a different version is specified through `$RUBY`.

Despite using the `shell` module, it appears the version of Ruby used by
Ansible is determined when `ansible-playbook` is invoked. By setting
`$RUBY` outside of Ansible, the correct version is used by the playbook.

The end result has a positive twist. A lot of previous commits have been
done to appease Travis CI. This change results in the version of Ruby
being manipulated only when running Travis CI's script.