Only download and install gitsh when needed
Using package managers to install libraries makes it really easy for
Ansible to skip unnecessary actions. When downloading, extracting, and
installing libraries through individual tasks, some ingenuity is needed
to skip actions that aren't needed. By checking for the version of
gitsh, these steps can be skipped when the correct version is already
installed.

This should help with the idempotence test on Travis CI.